+++ R/ggmice.R
@@ -3,11 +3,19 @@
 #' @param data An incomplete dataset (of class `data.frame`), or an object of class [`mice::mids`].
 #' @param mapping A list of aesthetic mappings created with [ggplot2::aes()].
 #'
-#' @return An object of class [`ggplot2::ggplot`].
+#' @return An object of class [`ggplot2::ggplot`]. The [`ggmice::ggmice`] function returns output
+#' equivalent to [`ggplot2::ggplot`] output, with a few important exceptions:
+#'
+#' - The theme is set to [`ggmice::theme_mice`].
+#' - The color scale is set to the [`mice::mdc`] colors.
+#' - The `colour` aesthetic is set to `.where`, an internally defined variable which distinguishes
+#' observed data from missing data or imputed data (for incomplete and imputed data, respectively).
 #'
 #' @examples
 #' dat <- mice::nhanes
 #' ggmice(dat, ggplot2::aes(x = age, y = bmi)) + ggplot2::geom_point()
+#' imp <- mice::mice(dat, print = FALSE)
+#' ggmice(imp, ggplot2::aes(x = age, y = bmi)) + ggplot2::geom_point()
 #' @seealso See the `ggmice` vignette to use the `ggmice()` function on
 #' [incomplete data](https://amices.org/ggmice/articles/ggmice.html#the-ggmice-function)
 #' or [imputed data](https://amices.org/ggmice/articles/ggmice.html#the-ggmice-function-1).
@@ -52,8 +60,7 @@ ggmice <- function(data = NULL,
   }
   if (length(vrbs) > length(unique(vrbs))) {
     cli::cli_abort(
-      c("The data must have unique column names.",
-        "x" = "Duplication found in {vrbs[duplicated(vrbs)]}")
+      c("The data must have unique column names.", "x" = "Duplication found in {vrbs[duplicated(vrbs)]}")
     )
   }
   # extract mapping variables
@@ -146,7 +153,6 @@ ggmice <- function(data = NULL,
   return(gg)
 }
 
-
 #' Utils function to extract mapping variables
 #'
 #' @param data Incomplete dataset or mids object.
@@ -189,11 +195,9 @@ match_mapping <- function(data, vrbs, mapping_in) {
         inherits(try(dplyr::mutate(mapping_data,
                                    !!rlang::parse_quo(mapping_text, env = rlang::current_env())),
                      silent = TRUE)
-                 ,
-                 "try-error")) {
+                 , "try-error")) {
       cli::cli_abort(
-        c("Must provide a valid mapping variable.",
-          "x" = "Mapping variable '{mapping_text}' not found in the data or imputations.")
+        c("Must provide a valid mapping variable.", "x" = "Mapping variable '{mapping_text}' not found in the data or imputations.")
       )
     } else {
       cli::cli_warn(