+++ R/plot_corr.R
@@ -11,7 +11,22 @@
 #' @return An object of class [ggplot2::ggplot].
 #'
 #' @examples
-#' plot_corr(mice::nhanes, label = TRUE)
+#' # plot correlations for all columns
+#' plot_corr(mice::nhanes)
+#'
+#' # plot correlations for specific columns by supplying a character vector
+#' plot_corr(mice::nhanes, c("chl", "hyp"))
+#'
+#' # plot correlations for specific columns by supplying unquoted variable names
+#' plot_corr(mice::nhanes, c(chl, hyp))
+#'
+#' # plot correlations for specific columns by passing an object with variable names
+#' # from the environment, unquoted with `!!`
+#' my_variables <- c("chl", "hyp")
+#' plot_corr(mice::nhanes, !!my_variables)
+#' # object with variable names must be unquoted with `!!`
+#' try(plot_corr(mice::nhanes, my_variables))
+#'
 #' @export
 plot_corr <-
   function(data,
@@ -25,38 +40,33 @@ plot_corr <-
       data <- as.data.frame(data)
     }
     verify_data(data = data, df = TRUE)
-    vrb <- substitute(vrb)
-    if (vrb != "all" && length(vrb) < 2) {
+    vrb <- rlang::enexpr(vrb)
+    vrb_matched <- match_vrb(vrb, names(data))
+    if (length(vrb_matched) < 2) {
       cli::cli_abort("The number of variables should be two or more to compute correlations.")
     }
-    if (vrb[1] == "all") {
-      vrb <- names(data)
-    } else {
-      data <- dplyr::select(data, {{vrb}})
-      vrb <- names(data)
-    }
     # check if any column is constant
-    constants <- apply(data, MARGIN = 2, function(x) {
+    constants <- apply(data[, vrb_matched], MARGIN = 2, function(x) {
       all(is.na(x)) || max(x, na.rm = TRUE) == min(x, na.rm = TRUE)
     })
     if (any(constants)) {
-      vrb <- names(data[, !constants])
+      vrb_matched <- vrb_matched[!constants]
       cli::cli_inform(
         c(
           "No correlations computed for variable(s):",
           " " = paste(names(constants[which(constants)]), collapse = ", "),
-          "x" = "Correlation undefined for constants."
+          "i" = "Correlations are undefined for constants."
         )
       )
     }
-
-    p <- length(vrb)
+    # compute correlations
+    p <- length(vrb_matched)
     corrs <- data.frame(
-      vrb = rep(vrb, each = p),
-      prd = vrb,
+      vrb = rep(vrb_matched, each = p),
+      prd = vrb_matched,
       corr = matrix(
         round(stats::cov2cor(
-          stats::cov(data.matrix(data[, vrb]), use = "pairwise.complete.obs")
+          stats::cov(data.matrix(data[, vrb_matched]), use = "pairwise.complete.obs")
         ), 2),
         nrow = p * p,
         byrow = TRUE
@@ -65,6 +75,7 @@ plot_corr <-
     if (!diagonal) {
       corrs[corrs$vrb == corrs$prd, "corr"] <- NA
     }
+    # create plot
     gg <-
       ggplot2::ggplot(corrs,
                       ggplot2::aes(
@@ -74,8 +85,8 @@ plot_corr <-
                         fill = .data$corr
                       )) +
       ggplot2::geom_tile(color = "black", alpha = 0.6) +
-      ggplot2::scale_x_discrete(limits = vrb, position = "top") +
-      ggplot2::scale_y_discrete(limits = rev(vrb)) +
+      ggplot2::scale_x_discrete(limits = vrb_matched, position = "top") +
+      ggplot2::scale_y_discrete(limits = rev(vrb_matched)) +
       ggplot2::scale_fill_gradient2(
         low = ggplot2::alpha("deepskyblue", 0.6),
         mid = "lightyellow",
@@ -91,13 +102,11 @@ plot_corr <-
           y = "Variable to impute",
           fill = "Correlation*
       ",
-      caption = "*pairwise complete observations"
+          caption = "*pairwise complete observations"
         )
     } else {
       gg <- gg +
-        ggplot2::labs(x = "Imputation model predictor",
-                      y = "Variable to impute",
-                      fill = "Correlation")
+        ggplot2::labs(x = "Imputation model predictor", y = "Variable to impute", fill = "Correlation")
     }
     if (label) {
       gg <-